[K1] fix getset test
	* gcc.target/k1/getset.c (pcr): Change to long so that the test
	works as expected in 64bits.

+++ gcc/testsuite/gcc.target/k1/getset.c
@@ -41,7 +41,7 @@ profile(unsigned int iter)
     int timer1 = 0;
     int timer2 = 0;
     int timer3 = 0;
-    unsigned int pcr;
+    unsigned long pcr;
 
     for (i = 0; i < iter; ++i) {
         TH_StartTimer ();
